made final changes for to script so it can handle processing of PTAB data in addition to Grants and Pubs
